// template/src/types/imageUpload.ts
export interface ImageUploadOptions {
  creationDate: string;
  cropRect: any;
  data: any;
  duration: any;
  exif: any;
  filename: string;
  height: number;
  localIdentifier: string;
  mime: string;
  modificationDate: string;
  path: string;
  size: number;
  sourceURL: string;
  width: number;
}